@@ -11,7 +11,7 @@ import 'product_title_with_image.dart';
 class Body extends StatelessWidget {
   final Product product;
 
-  const Body({Key key, this.product}) : super(key: key);
+  const Body({Key? key, required this.product}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     // It provide us total height and width
@@ -19,40 +19,37 @@ class Body extends StatelessWidget {
     return SingleChildScrollView(
       child: Column(
         children: <Widget>[
-          SizedBox(
-            height: size.height,
-            child: Stack(
-              children: <Widget>[
-                Container(
-                  margin: EdgeInsets.only(top: size.height * 0.3),
-                  padding: EdgeInsets.only(
-                    top: size.height * 0.12,
-                    left: kDefaultPaddin,
-                    right: kDefaultPaddin,
-                  ),
-                  // height: 500,
-                  decoration: BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.only(
-                      topLeft: Radius.circular(24),
-                      topRight: Radius.circular(24),
-                    ),
-                  ),
-                  child: Column(
-                    children: <Widget>[
-                      ColorAndSize(product: product),
-                      SizedBox(height: kDefaultPaddin / 2),
-                      Description(product: product),
-                      SizedBox(height: kDefaultPaddin / 2),
-                      CounterWithFavBtn(),
-                      SizedBox(height: kDefaultPaddin / 2),
-                      AddToCart(product: product)
-                    ],
+          Stack(
+            children: <Widget>[
+              Container(
+                margin: EdgeInsets.only(top: size.height * 0.35),
+                padding: EdgeInsets.only(
+                  top: size.height * 0.08,
+                  left: kDefaultPaddin,
+                  right: kDefaultPaddin,
+                ),
+                // height: 500,
+                decoration: BoxDecoration(
+                  color: Colors.white,
+                  borderRadius: BorderRadius.only(
+                    topLeft: Radius.circular(24),
+                    topRight: Radius.circular(24),
                   ),
                 ),
-                ProductTitleWithImage(product: product)
-              ],
-            ),
+                child: Column(
+                  children: <Widget>[
+                    ColorAndSize(product: product),
+                    SizedBox(height: kDefaultPaddin / 2),
+                    Description(product: product),
+                    SizedBox(height: kDefaultPaddin / 2),
+                    CounterWithFavBtn(),
+                    SizedBox(height: kDefaultPaddin / 2),
+                    AddToCart(product: product)
+                  ],
+                ),
+              ),
+              ProductTitleWithImage(product: product)
+            ],
           )
         ],
       ),
